+++ prototyping_inference_engine/unifier/disjunctive_piece_unifier_algorithm.py
@@ -12,16 +12,16 @@ from prototyping_inference_engine.api.ontology.rule.rule import Rule
 from prototyping_inference_engine.api.ontology.rule.validators import ensure_ed_rule
 from prototyping_inference_engine.api.query.conjunctive_query import ConjunctiveQuery
 from prototyping_inference_engine.api.query.union_query import UnionQuery
-from prototyping_inference_engine.backward_chaining.unifier.disjunctive_piece_unifier import (
+from prototyping_inference_engine.unifier.disjunctive_piece_unifier import (
     DisjunctivePieceUnifier,
 )
-from prototyping_inference_engine.backward_chaining.unifier.piece_unifier import (
+from prototyping_inference_engine.unifier.piece_unifier import (
     PieceUnifier,
 )
-from prototyping_inference_engine.backward_chaining.unifier.piece_unifier_algorithm import (
+from prototyping_inference_engine.unifier.piece_unifier_algorithm import (
     PieceUnifierAlgorithm,
 )
-from prototyping_inference_engine.backward_chaining.unifier.piece_unifier_cache import (
+from prototyping_inference_engine.unifier.piece_unifier_cache import (
     PieceUnifierCache,
 )