+++ prototyping_inference_engine/unifier/piece_unifier_cache.py
@@ -11,7 +11,7 @@ from typing import Optional, Iterable
 from prototyping_inference_engine.api.atom.term.term import Term
 from prototyping_inference_engine.api.ontology.rule.rule import Rule
 from prototyping_inference_engine.api.query.conjunctive_query import ConjunctiveQuery
-from prototyping_inference_engine.backward_chaining.unifier.piece_unifier import (
+from prototyping_inference_engine.unifier.piece_unifier import (
     PieceUnifier,
 )